+++ src/config.js
@@ -3,9 +3,8 @@
 
     window.LampaYani = window.LampaYani || {};
     window.LampaYani.Config = window.LampaYaniConfig = {
-        version: '0.20.14',
+        version: '0.20.15',
         apiBase: 'https://api.yani.tv',
-        episodesApiBase: 'https://yummytv.kemonos.win/api',
         statusUrl: 'https://andrewcodeman.github.io/lampa_yani/status/status.json',
         applicationHeader: 'p6_gpujl6d3pho8n', // Public Yani application token
         cacheTtl: 300000,